// src/xpanse-api/generated/models/FlavorBasic.ts
/*
 * SPDX-License-Identifier: Apache-2.0
 * SPDX-FileCopyrightText: Huawei Inc.
 */

/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * The flavors of the orderable service.
 */
export type FlavorBasic = {
    /**
     * The flavor name
     */
    name: string;
    /**
     * The price of the flavor
     */
    fixedPrice: number;
};